requirements.txt
mypy
fastapi[standard]
uvicorn[standard]
psycopg2-binary
python-dotenv
pytest
pytest-asyncio
pytest-mock
pytest-cov